+++ src/util/inventory.py
@@ -1,0 +1,11 @@
+
+
+class InventoryUtils:
+    """
+    A class of inventory utilities, functions return false on error
+    """
+
+    @staticmethod
+    def add() -> bool:
+        pass
+
